Rename ifsbench.launcher to ifsbench.launch
