+++ ifsbench/__init__.py
@@ -25,7 +25,7 @@ from ifsbench.files import * # noqa
 from ifsbench.gribfile import * # noqa
 from ifsbench.ifs import * # noqa
 from ifsbench.job import * # noqa
-from ifsbench.launcher import * # noqa
+from ifsbench.launch import * # noqa
 from ifsbench.logging import * # noqa
 from ifsbench.namelist import * # noqa
 from ifsbench.nodefile import * # noqa
